requirements.txt
gunicorn==21.2.0
uvicorn==0.23.2
fastapi==0.103.1
pydantic==2.4.2
pydantic_core==2.10.1
mysql.connector==2.2.9